2411013 게시판 구현 검색어 저장 시작 상품 이미지 업로드 여전히 안됨

+++ src/main/java/com/example/byosinsa/dto/BoardDTO.java
@@ -19,12 +19,12 @@ public class BoardDTO {
 
     private String bWriter;
 
-    private Long cno;
+    private int cno;
 
-    private Long cName;
+    private int cName;
 
     // 조회수
-    private Long views;
+    private int views;
 
 
     @Builder.Default
