+++ src/main/java/com/example/byosinsa/repository/BoardCategoryRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BoardCategoryRepository extends JpaRepository<BoardCategory, Long> {
+public interface BoardCategoryRepository extends JpaRepository<BoardCategory, Integer> {
 
 }
